backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0           # Async PostgreSQL driver
alembic==1.13.1

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0  # JWT encoding/decoding
passlib[bcrypt]==1.7.4            # Password hashing
bcrypt==4.1.2

# HTTP Client (for USDA API)
httpx==0.26.0             # Async HTTP client

# Caching
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0             # Also used for test client

# Development
python-dotenv==1.0.0
